+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - **operator**: add `element_at` operator.
 - **operator**: add `ignore_elements` operator.
 - **operator**: add `all` operator.
+- **operator**: add `contains` operator.
 
 ### Refactor
 - **operator**: `skip_last` should emit not only when observer complete.